Exams continued (#485)
* Add exam timer

* Exam page progress

* Fixed incorrect type on insert

* Add initial exam start banner

* Progress on the exam feature

* Accessibility for multiple choice (#495)

* Accessibility for multiple choice

* update snapshots

* Fixed not taking possible null into account

* Improved exam grading

* Return grading only when exam is over

* Fix return type and regenerate bindings

* Multiple choice grading view (#496)

* Exam over notification

* Initial grading view for quizzes multiple choice

* Fixes

* Style tweaks etc.

* Fixes

* Support Markdown on buttons

* Various improvements

* Progress

* Added missing authorize calls

* Grey selected answers in grading, update snapshots

* Include model solution spec for ended exams

* Fixed eslint errors

+++ services/headless-lms/src/controllers/main_frontend/exercise_services.rs
@@ -1,5 +1,6 @@
 //! Controllers for requests starting with `/api/v0/main-frontend/exercise-services/`.
 
+use crate::domain::authorization::{authorize, Action, Resource};
 use crate::models::exercise_services::ExerciseServiceNewOrUpdate;
 use crate::{
     controllers::ControllerResult, domain::authorization::AuthUser,
@@ -17,6 +18,8 @@ async fn delete_exercise_service(
     user: AuthUser,
 ) -> ControllerResult<Json<ExerciseService>> {
     let mut conn = pool.acquire().await?;
+    authorize(&mut conn, Action::Edit, user.id, Resource::ExerciseService).await?;
+
     let deleted = crate::models::exercise_services::delete_exercise_service(
         &mut conn,
         *request_exercise_service_id,
@@ -32,6 +35,8 @@ async fn add_exercise_service(
     payload: web::Json<ExerciseServiceNewOrUpdate>,
 ) -> ControllerResult<Json<ExerciseService>> {
     let mut conn = pool.acquire().await?;
+    authorize(&mut conn, Action::Edit, user.id, Resource::ExerciseService).await?;
+
     let exercise_service = payload.0;
     let created =
         crate::models::exercise_services::insert_exercise_service(&mut conn, &exercise_service)
@@ -73,6 +78,8 @@ async fn update_exercise_service(
     user: AuthUser,
 ) -> ControllerResult<Json<ExerciseService>> {
     let mut conn = pool.acquire().await?;
+    authorize(&mut conn, Action::Edit, user.id, Resource::ExerciseService).await?;
+
     let updated_exercise_service = payload.0;
     let updated_service = crate::models::exercise_services::update_exercise_service(
         &mut conn,
